+++ apps/frontend/src/components/structure/framing/Native/NativeFrameManager.tsx
@@ -10,7 +10,7 @@
  * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React from 'react';
+import React, { useEffect } from 'react';
 import MailPage from '@/pages/Mail/MailPage';
 import useAppConfigsStore from '@/pages/Settings/AppConfig/appConfigsStore';
 import useFrameStore from '@/components/structure/framing/useFrameStore';
@@ -18,6 +18,10 @@ import LinuxmusterPage from '@/pages/LinuxmusterPage/LinuxmusterPage';
 import type AppConfigDto from '@libs/appconfig/types/appConfigDto';
 import APP_INTEGRATION_VARIANT from '@libs/appconfig/constants/appIntegrationVariants';
 import APPS from '@libs/appconfig/constants/apps';
+import { useLocation } from 'react-router-dom';
+import useUserStore from '@/store/UserStore/useUserStore';
+import findAppConfigByName from '@libs/common/utils/findAppConfigByName';
+import getRootPathName from '@libs/common/utils/getRootPathName';
 
 const isActiveNativeFrame = (appConfig: AppConfigDto, loadedFrames: string[]) => {
   const { appType } = appConfig;
@@ -26,8 +30,26 @@ const isActiveNativeFrame = (appConfig: AppConfigDto, loadedFrames: string[]) =>
 };
 
 const NativeFrameManager = () => {
+  const { pathname } = useLocation();
+  const rootPathName = getRootPathName(pathname);
+
   const { appConfigs } = useAppConfigsStore();
-  const { loadedEmbeddedFrames } = useFrameStore();
+  const { isAuthenticated } = useUserStore();
+  const { setEmbeddedFrameLoaded, setActiveEmbeddedFrame, loadedEmbeddedFrames } = useFrameStore();
+
+  useEffect(() => {
+    const appName = findAppConfigByName(appConfigs, rootPathName)?.name;
+    if (isAuthenticated && appName) {
+      setEmbeddedFrameLoaded(appName);
+      setActiveEmbeddedFrame(appName);
+    } else {
+      setActiveEmbeddedFrame(null);
+    }
+
+    return () => {
+      setActiveEmbeddedFrame(null);
+    };
+  }, [isAuthenticated, pathname]);
 
   return appConfigs
     .filter((appConfig) => isActiveNativeFrame(appConfig, loadedEmbeddedFrames))
